+++ src/main.rs
@@ -6,6 +6,28 @@ use iced::{window, Application, Settings};
 use ytdlp_gui::{logging, Config, YtGUI};
 
 fn main() -> iced::Result {
+    let mut args = std::env::args();
+
+    // don't need first arg (the executable path)
+    args.next();
+
+    if let Some(first_arg) = args.next() {
+        if first_arg == "--help" || first_arg == "-h" {
+            println!("Usage: ytdlp-gui <OPTIONS>\n");
+            println!("Options:");
+            println!("-h, --help     Print help");
+            println!("-V, --version  Print version");
+            std::process::exit(0);
+        } else if first_arg == "--version" || first_arg == "-V" {
+            let version = option_env!("CARGO_PKG_VERSION").unwrap_or("unknown");
+            println!("version: {version}");
+            std::process::exit(0);
+        } else {
+            println!("Invalid option/argument");
+            std::process::exit(1);
+        }
+    }
+
     logging();
 
     let config_dir = dirs::config_dir()