+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## Unversioned
+## 1.0.0
+(got bored of the 0.whatever so I'm bumping it to 1.0.0)
 
 - Major: Better download progress messages (including progress for playlist videos) since there is better yt-dlp parsing
-- Major: Move downloads logs from `config_dir` to `cache_dir` linux: $XDG_CACHE_HOME or $HOME/.cache -- windows: {FOLDERID_LocalAppData} -- macos: $HOME/Library/Caches
+- Major: Move downloads logs from `config_dir` to `cache_dir` linux: `$XDG_CACHE_HOME` or `$HOME/.cache` -- windows: `{FOLDERID_LocalAppData}` -- macos: `$HOME/Library/Caches`
 - Major: Dev: Better yt-dlp parsing
 - Bugfix: Use `.to_string_lossy()` for download dir instead of `.to_str()`, that will solve [this issue](https://github.com/BKSalman/ytdlp-gui/issues/12)
 - Minor: Use default configs if config file is broken